# src/main/resources/static/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: Yamb API
  description: API for playing Yamb.
  version: 1.0.1
...
paths:
...
  /api/players:
    get:
      summary: Get all players
      tags:
        - Players
      responses:
        '200':
          description: List of all players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerResponse'
        '500':
          description: Internal server error
  /api/players/{id}:
    get:
      summary: Get player by ID
      tags:
        - Players
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the player
          schema:
            type: integer
      responses:
        '200':
          description: Player details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerResponse'
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  /api/players/{id}/stats:
    get:
      summary: Get player stats by ID
      tags:
        - Players
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the player
          schema:
            type: integer
      responses:
        '200':
          description: Player statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api/playerstatsResponse'
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  /api/players/{id}/scores:
    get:
      summary: Get player scores by ID
      tags:
        - Players
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the player
          schema:
            type: integer
      responses:
        '200':
          description: List of scores for the player
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShortScoreResponse'
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  /api/players/{id}/preferences:
    get:
      summary: Get player preferences by ID
      tags:
        - Players
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the player
          schema:
            type: integer
      responses:
        '200':
          description: Player preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerPreferencesResponse'
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error

    put:
      summary: Set player preferences by ID
      tags:
        - Players
      security:
        - bearerAuth: []
      requestBody:
        description: Player preferences to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerPreferencesRequest'
      responses:
        '200':
          description: Updated player preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerPreferencesResponse'
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  /api/scores:
    get:
      summary: Get all scores
      tags:
        - Scores
      responses:
        '200':
          description: List of all scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreResponse'
        '500':
          description: Internal server error
  /api/scores/{id}:
    get:
      summary: Get score by ID
      tags:
        - Scores
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the score
          schema:
            type: integer
      responses:
        '200':
          description: Score details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreResponse'
        '404':
          description: Score not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  /api/scores/stats:
    get:
      summary: Get global score statistics
      tags:
        - Scores
      responses:
        '200':
          description: Global score statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalScoreStatsResponse'
        '500':
          description: Internal server error
...